% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papaya_version.R
\name{get_papaya_version}
\alias{get_papaya_version}
\title{Get Papaya Version}
\usage{
get_papaya_version()
}
\value{
List of build and version, both characters
}
\description{
Reads the papaya.js file installed and determines version and build
}

