% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/papaya.R
\name{papaya}
\alias{papaya}
\title{View images with Papaya}
\usage{
papaya(images)
}
\arguments{
\item{images}{character filenames or \code{nifti} objects to be viewed}
}
\value{
Output directory where index.html, js, and copied nii.gz files
}
\description{
Writes temporary images out from nifti objects or passes character filenames
of images to papaya JS viewer
}

