% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papaya_div.R
\name{papaya_div}
\alias{papaya_div}
\title{Papaya Div element output}
\usage{
papaya_div()
}
\value{
Character string
}
\description{
Get the necessary div output for embedding a papaya image
}
\examples{
papaya_div()
}

