% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papaya_version.R
\name{update_papaya_build}
\alias{update_papaya_build}
\title{Update Papaya build version from GitHub}
\usage{
update_papaya_build(type = c("standard", "minimal", "nodicom", "nojquery",
  "standard-with-atlas-local", "standard-with-atlas"), verbose = TRUE)
}
\arguments{
\item{type}{Type of release.  Standard is default}

\item{verbose}{Should download progress be shown?}
}
\value{
Result of \code{\link{get_papaya_version}} after downloading
}
\description{
Updates the papaya version in the papayar package to the most 
current on GitHub
}

