% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/papaya.R
\name{papaya}
\alias{papaya}
\title{View images with Papaya}
\usage{
papaya(images)
}
\arguments{
\item{images}{character filenames or \code{nifti} objects to be viewed}
}
\value{
Output directory where index.html, js, and copied nii.gz files
}
\description{
Writes temporary images out from nifti objects or passes character filenames
of images to papaya JS viewer
}
\examples{
x = nifti(img = array(rnorm(100^3), dim= rep(100, 3)), dim=rep(100, 3), datatype=16)
thresh = datatyper(x > 1)
odir = papaya(list(x, thresh))
cat(paste0("# All files are located in \\n", odir))
}

