% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pass_papaya.R
\name{pass_papaya}
\alias{pass_papaya}
\title{View images with Papaya}
\usage{
pass_papaya(L = NULL, outdir = NULL, daemon = FALSE,
  close_on_exit = TRUE, sleeper = 3)
}
\arguments{
\item{L}{list of arguments passed to papaya using params}

\item{outdir}{output directory for index and all to go}

\item{daemon}{Argument passed to \code{\link[servr]{server_config}}}

\item{close_on_exit}{Should the server close once the function finishes?}

\item{sleeper}{Time in seconds to sleep if \code{close_on_exit = TRUE}.
This allows the server to start up.}
}
\description{
Writes temporary images out from nifti objects or passes character filenames
of images to papaya JS viewer
}

